fix: 캘린더 페이지 내 함수 유틸로 분리
