+++ lib/views/plan/plan/calendar/calendar_page.dart
@@ -1,30 +1,12 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:travel_muse_app/providers/plan/calendar_provider.dart';
+import 'package:travel_muse_app/utills/date_utils.dart';
 import 'package:travel_muse_app/views/plan/plan/calendar/widgets/calendar_widget.dart';
 import 'package:travel_muse_app/views/plan/plan/location_setting/province_setting_page.dart';
 
 class CalendarPage extends ConsumerWidget {
   const CalendarPage({super.key});
-
-  // 여행 일수 계산 함수
-  int getTripDays(DateTime start, DateTime end) {
-    return end.difference(start).inDays + 1;
-  }
-
-  // 버튼에 보여줄 텍스트 생성 함수
-  String getButtonText(DateTime? startDay, DateTime? endDay) {
-    if (startDay == null || endDay == null) {
-      return '다음';
-    } else {
-      final startStr =
-          '${startDay.year}.${startDay.month.toString().padLeft(2, '0')}.${startDay.day.toString().padLeft(2, '0')}';
-      final endStr =
-          '${endDay.year}.${endDay.month.toString().padLeft(2, '0')}.${endDay.day.toString().padLeft(2, '0')}';
-      final days = getTripDays(startDay, endDay);
-      return '$startStr ~ $endStr ($days일) 선택하기';
-    }
-  }
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
